add: Docker configs

+++ microservicios/tata-test-service/test-tata-dataaccess-service/src/main/java/tata/test/dataaccess/entities/CuentaEntity.java
@@ -48,7 +48,7 @@ public class CuentaEntity extends AbstractEntity<Integer, Integer> {
   @Column(name = "saldoInicial", nullable = false)
   @NotNull(message = "Saldo inicial no debe ser nulo")
   @PositiveOrZero(message = "Saldo inicial debe ser positivo o cero")
-  private Integer saldoInicial;
+  private double saldoInicial;
 
   @ManyToOne(fetch = FetchType.EAGER)
   @JoinColumn(name = "identificacion", nullable = false)
