+++ microservicios/tata-test-service/test-tata-domain-service/test-tata-domain-core-service/src/main/java/tata/test/record/request/CuentaRequestRecord.java
@@ -23,7 +23,7 @@ public record CuentaRequestRecord(
     Integer id,
     String numeroCuenta,
     String tipoCuenta,
-    Integer saldoInicial,
+    double saldoInicial,
     String estado,
     String cedulaCliente
 ) {
