+++ microservicios/tata-test-service/test-tata-domain-service/test-tata-domain-core-service/src/main/java/tata/test/record/response/CuentaResponseRecord.java
@@ -25,7 +25,7 @@ public record CuentaResponseRecord(
     Integer id,
     String numeroCuenta,
     String tipoCuenta,
-    Integer saldoInicial,
+    double saldoInicial,
     @JsonIgnore String estado,
     @JsonIgnore Integer idPersonaCrea,
     @JsonIgnore LocalDateTime fechaCrea,
